backend/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
gunicorn==20.1.0
flask-pymongo==2.3.0
flask-cors==3.0.10
PyJWT==2.8.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
bcrypt==4.1.2
